Extract watch communication message models into HAWatchCommunicationMessages package
Move the WatchConnectivity wire-crossing identifier types out of Shared into a
new dependency-free local SPM package, one type per file:

- InteractiveImmediateMessages / InteractiveImmediateResponses
- GuaranteedMessages
- MagicItemExecutionFailureCode
- WatchContext (extracted from WatchHelpers.swift, now documenting which
  device owns each application-context key)

The package is linked by Shared-iOS/Shared-watchOS only (mirroring HAModels)
and re-exported via HADesignSystemExports.swift, so all existing call sites
compile unchanged. Payload models (WatchConfig, WatchDatabaseMirror,
WatchConfigAvailableItems, MagicItem.Info) stay in Shared until MagicItem can
be extracted from Shared.
