--- Sources/Shared/Watch/GuaranteedMessages.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public enum GuaranteedMessages: String, CaseIterable {
-    case sync
-}
